Add files via upload
Writing a program that asks for a number N as input. After that, the program will keep asking for an integer as input until the integer being sent as input is equal to N, and at that point, the program will print “Congratulations! Your input is equal to N!!” and quit. During the repeated queries, if the number you entered was less than N, print “Your input is less than N” and if the number was greater than N, print “Your input is greater than N”.