Add files via upload
i) Takes as input an array of N names, an array of N ids and an array of N salaries where the 
name, id, and salary of the i
th index belongs to the same person. 
ii) Stores all the above information in an array of structures. Then, prints “Information Stored 
Successfully.”
iii) Initiates a search operation: Takes a name of a person from the user (See Intermediate 
Output) to search from the array of persons. If found, displays the information of the 
requested person (See Output). If the person‟s salary is more than 16000, the status of the 
person is “A” or else the status is “B”.